[change] download only if file does not found

+++ init.sh
@@ -1,6 +1,7 @@
 #!/usr/bin/env bash
 
-# deployment
+# deploy dotfiles
+
 mkdir -p ~/.hammerspoon
 mkdir -p ~/vimbackup
 mkdir -p ~/.vim
@@ -21,14 +22,29 @@ ln -snf ~/dotfiles/colors ~/.vim/colors
 ln -snf ~/dotfiles/.docker/config.json ~/.docker/config.json
 
 
-# init
-curl -L -o ~/.git-completion.bash https://raw.github.com/git/git/master/contrib/completion/git-completion.bash
-curl -L -o ~/.git-prompt.sh https://github.com/git/git/raw/master/contrib/completion/git-prompt.sh
-chmod +x ~/.git-prompt.sh
+# bash git completion
+
+if [ ! -f ~/.git-completion.bash ]; then
+  curl -L -o ~/.git-completion.bash https://raw.github.com/git/git/master/contrib/completion/git-completion.bash
+fi
+if [ ! -f ~/.git-prompt.sh ]; then
+  curl -L -o ~/.git-prompt.sh https://github.com/git/git/raw/master/contrib/completion/git-prompt.sh
+  chmod +x ~/.git-prompt.sh
+fi
+
+
+# zsh git completion
+
 ZSH_COMPLETION_PATH=~/.zsh/completion
 mkdir -p $ZSH_COMPLETION_PATH
-curl -L -o $ZSH_COMPLETION_PATH/git-completion.bash https://raw.github.com/git/git/master/contrib/completion/git-completion.bash
-curl -L -o $ZSH_COMPLETION_PATH/_git https://raw.github.com/git/git/master/contrib/completion/git-completion.zsh
-curl -L -o $ZSH_COMPLETION_PATH/git-prompt.sh https://github.com/git/git/raw/master/contrib/completion/git-prompt.sh
+if [ ! -f $ZSH_COMPLETION_PATH/git-completion.bash ]; then
+  curl -L -o $ZSH_COMPLETION_PATH/git-completion.bash https://raw.github.com/git/git/master/contrib/completion/git-completion.bash
+fi
+if [ ! -f $ZSH_COMPLETION_PATH/_git ]; then
+  curl -L -o $ZSH_COMPLETION_PATH/_git https://raw.github.com/git/git/master/contrib/completion/git-completion.zsh
+fi
+if [ ! -f $ZSH_COMPLETION_PATH/git-prompt.sh ]; then
+  curl -L -o $ZSH_COMPLETION_PATH/git-prompt.sh https://github.com/git/git/raw/master/contrib/completion/git-prompt.sh
+fi
 rm -f ~/.zcompdump
 which compinit && compinit
